+++ OnlineStore/OnlineStore.Application/DTOs/CategoryDTOs/CategoryDto.cs
@@ -1,0 +1,7 @@
+﻿namespace OnlineStore.Application.DTOs.CategoryDTOs;
+
+public class CategoryDto
+{
+    public long Id { get; set; }
+    public string Name { get; set; }
+}
