+++ OnlineStore/OnlineStore.Application/DTOs/CategoryDTOs/CreateCategoryDto.cs
@@ -1,0 +1,5 @@
+﻿namespace OnlineStore.Application.DTOs.CategoryDTOs;
+public class CreateCategoryDto
+{
+    public string Name { get; set; }
+}
