+++ OnlineStore/OnlineStore.Application/DTOs/CategoryDTOs/UpdateCategoryDto.cs
@@ -1,0 +1,7 @@
+﻿namespace OnlineStore.Application.DTOs;
+
+public class UpdateCategoryDto
+{
+    public long Id { get; set; }
+    public string Name { get; set; }
+}
