+++ OnlineStore/OnlineStore.Application/DTOs/OrderDTOs/CreateOrderDto.cs
@@ -1,0 +1,5 @@
+﻿namespace OnlineStore.Application.DTOs.OrderDTOs;
+public class CreateOrderDto
+{
+    public long UserId { get; set; }
+}
