+++ OnlineStore/OnlineStore.Application/DTOs/OrderDTOs/OrderDto.cs
@@ -1,0 +1,8 @@
+﻿namespace OnlineStore.Application.DTOs.OrderDTOs;
+
+public class OrderDto
+{
+    public long Id { get; set; }
+    public DateTime CreatedAt { get; set; }
+    public long UserId { get; set; }
+}
