+++ OnlineStore/OnlineStore.Application/DTOs/ProductDTOs/CreateProductDto.cs
@@ -1,0 +1,9 @@
+﻿namespace OnlineStore.Application.DTOs.ProductDTOs;
+
+public class CreateProductDto
+{
+    public string Name { get; set; }
+    public string Description { get; set; }
+    public double Price { get; set; }
+    public long CategoryId { get; set; }
+}
