+++ OnlineStore/OnlineStore.Application/DTOs/ProductDTOs/ProductDto.cs
@@ -1,0 +1,8 @@
+﻿namespace OnlineStore.Application.DTOs.ProductDTOs;
+
+public class ProductDto
+{
+    public long Id { get; set; }
+    public string Name { get; set; }
+    public double Price { get; set; }
+}
