+++ OnlineStore/OnlineStore.Application/DependencyInjection/DependencyInjection.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.Extensions.DependencyInjection;
 using OnlineStore.Application.Mapping;
+using System.Reflection;
 
 namespace OnlineStore.Application.DependencyInjection;
 
@@ -8,13 +9,11 @@ public static class DependencyInjection
 {
     public static void AddApplicationLayer(this IServiceCollection services)
     {
+        services.AddAutoMapper(Assembly.GetExecutingAssembly());
         InitServices(services);
     }
 
     private static void InitServices(this IServiceCollection services)
     {
-        services.AddAutoMapper(typeof(CategoryProfile), 
-                               typeof(ProductProfile),
-                               typeof(OrderProfile));
     }
 }
