+++ OnlineStore/OnlineStore.Application/Mapping/CategoryProfile.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using OnlineStore.Application.DTOs;
+using OnlineStore.Application.DTOs.CategoryDTOs;
 using OnlineStore.Domain.Entities;
 
 namespace OnlineStore.Application.Mapping;
