+++ OnlineStore/OnlineStore.Application/Mapping/ProductProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using OnlineStore.Application.DTOs;
+using OnlineStore.Application.DTOs.ProductDTOs;
 using OnlineStore.Domain.Entities;
 
 namespace OnlineStore.Application.Mapping;
